Severe: added db migrations and seed for site_meta. Tried to set up a PageController to display page with site_meta

+++ database/migrations/2014_10_12_100000_create_password_resets_table.php
@@ -13,11 +13,13 @@ class CreatePasswordResetsTable extends Migration
      */
     public function up()
     {
-        Schema::create('password_resets', function (Blueprint $table) {
-            $table->string('email')->index();
-            $table->string('token')->index();
-            $table->timestamp('created_at')->nullable();
-        });
+        if(!Schema::hasTable('password_resets')) {
+            Schema::create('password_resets', function (Blueprint $table) {
+                $table->string('email')->index();
+                $table->string('token')->index();
+                $table->timestamp('created_at')->nullable();
+            });
+        }
     }
 
     /**
@@ -27,6 +29,8 @@ class CreatePasswordResetsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('password_resets');
+        if(Schema::hasTable('password_resets')) {
+            Schema::drop('password_resets');
+        }
     }
 }
